+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ## v0.8.3-dev
+* Enhancements
+  * add nested paramer type `Json`
 
 ## v0.8.2 (2015-9-26)
 * Documentation